requirements.txt
aiobotocore==2.4.0
aiohttp==3.8.1
aioitertools==0.11.0
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
boto3
botocore
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
elastic-transport==8.4.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
Faker==14.2.0
fastapi==0.68.2
frozenlist==1.3.1
fsspec==2022.8.2
h11==0.13.0
idna==3.3
jmespath==1.0.1
multidict==6.0.2
numpy==1.23.3
pandas==1.4.4
pycodestyle==2.9.1
pydantic==1.9.2
pydantic-factories==1.6.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
s3fs==2022.8.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.15.0
wrapt==1.14.1
xeger==0.3.5
yarl==1.8.1
